1122-02ver

+++ src/main/webapp/admin/adminBoardDeletePro.jsp
@@ -11,51 +11,78 @@
 <jsp:setProperty property="num" name="dto" />
 <jsp:useBean class="team03.bean.LocalBoardDTO" id="Ldto" />
 <jsp:setProperty property="num" name="Ldto" />
+<jsp:useBean class="team03.bean.MovieDTO" id="Mdto" />
+<jsp:setProperty property="num" name="Mdto" />
+
 
 <%
-	int num = Integer.parseInt(request.getParameter("num"));
-	int Lnum = Integer.parseInt(request.getParameter("Lnum"));
-	String pageNum = request.getParameter("pageNum");	
+	String num = request.getParameter("num");
+	String Lnum = request.getParameter("Lnum");
+	String Mnum = request.getParameter("Mnum");
 	
-	AdminDAO dao = AdminDAO.getInstance();
-	int result = dao.deleteContent(num);
-	int Lresult = dao.deleteLcontent(Lnum);
+	if(num != null){
+		
+		AdminDAO dao = AdminDAO.getInstance();
+		int result = dao.deleteContent(Integer.parseInt(num));
+		
+		if(result == 1){ 
+		String path = request.getRealPath("team03File");
+		File f = new File (path+"//"+result);
+		f.delete();%>
+
+			<script>
+				alert("삭제되었습니다.");
+				window.location="adminMain.jsp";
+			</script>
+		<%}else{%>
+			<script>
+				alert("오류.");
+				history.go(-1);
+			</script>
+		<%}
+	}
 	
-	if(num != 0){
-	if(result == 1){ 
+	if(Lnum != null){
+		
+		AdminDAO dao = AdminDAO.getInstance();
+		int result = dao.LocalDeleteContent(Integer.parseInt(Lnum));
+		
+		if(result == 1){
 			String path = request.getRealPath("team03File");
-			File f = new File(path+"//"+result);
-			f.delete();		%>	
+			File f = new File (path + "//" + result);
+			f.delete(); %>
 			
 			<script>
 				alert("삭제되었습니다.");
-				window.location="adminDeleteContentConfirm.jsp?pageNum=<%=pageNum%>";
+				window.location="adminMain.jsp";
 			</script>
-				
-	<%	} else { %>
+		<%}else{%>
 			<script>
-				alert("비밀번호가 틀립니다.");
+				alert("오류.");
 				history.go(-1);
 			</script>
-	<%	}%>
-	<% }
+		<%}
+	}
 	
-	if(Lnum != 0){
-	if(Lresult == 1){ 
-			Ldto.setNum(Lnum);
+	if(Mnum != null){
+		
+		AdminDAO dao = AdminDAO.getInstance();
+		int result = dao.MovieDeleteContent(Integer.parseInt(Mnum));
+		
+		if(result == 1){
 			String path = request.getRealPath("team03File");
-			File f = new File(path+"//"+result);
-			f.delete();		%>	
+			File f = new File(path + "//" + result);
+			f.delete(); %>
 			
 			<script>
 				alert("삭제되었습니다.");
-				window.location="adminDeleteContentConfirm.jsp?pageNum=<%=pageNum%>";
+				window.location="adminMain.jsp";
 			</script>
-				
-	<%	} else { %>
-			<script>
-				alert("비밀번호가 틀립니다.");
+		<%}else{%>
+				<script>
+				alert("오류.");
 				history.go(-1);
 			</script>
-	<%	}%>
-	<%} %>
+		<%}
+	}%>
+
